backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pillow==12.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.2
python-dotenv==1.2.1
cryptography